requirements.txt
aiohttp==3.6.2
alpha-vantage==2.2.0
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
DateTime==4.3
Django==3.0.7
django-cors==0.1
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.9
investpy==0.9.14
lxml==4.5.1
multidict==4.7.6
multitasking==0.0.9
numpy==1.19.0
pandas==1.0.5
pandas-datareader==0.8.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
TA-Lib==0.4.18
Unidecode==1.1.1
urllib3==1.25.9
yarl==1.4.2
yfinance==0.1.54
zope.interface==5.1.0